@@ -12,51 +12,52 @@
     "url": "https://github.com/idlist/monurls/issues"
   },
   "homepage": "https://github.com/idlist/monurls#readme",
+  "type": "module",
   "main": "backend/dist/index.js",
   "scripts": {
-    "dev": "vite",
-    "bundle": "vite build",
-    "build": "vite build && node backend/scripts/build",
-    "compile": "node backend/scripts/build",
-    "start": "node backend/scripts/build && node backend/dist/index",
-    "prod": "vite build && node backend/scripts/build && node backend/dist/index",
+    "f:dev": "vite",
+    "f:build": "vite build",
+    "b:build": "node backend/scripts/build",
     "host": "node backend/dist/index",
+    "start": "npm run b:build && npm run host",
+    "build": "npm run f:build && npm run b:build",
+    "prod": "npm run build && npm run host",
     "count": "cd .. && cloc monurls --exclude-dir=node_modules,dist --include-ext=ts,tsx,sass"
   },
   "dependencies": {
-    "axios": "^0.24.0",
-    "fastify": "^3.0.0",
-    "fastify-cookie": "^5.0.0",
-    "fastify-cors": "^6.0.0",
-    "fastify-plugin": "^3.0.0",
-    "fastify-rate-limit": "^5.0.0",
+    "@fastify/cookie": "^6.0.0",
+    "@fastify/cors": "^7.0.0",
+    "@fastify/rate-limit": "^6.0.1",
+    "@fastify/static": "^5.0.2",
+    "axios": "^0.27.2",
+    "fastify": "^3.29.0",
+    "fastify-plugin": "^3.0.1",
     "fastify-schedule": "^1.0.0",
-    "fastify-static": "^4.0.0",
-    "js-yaml": "^3.0.0",
-    "luxon": "^2.0.0",
-    "mariadb": "^2.0.0",
-    "normalize.css": "^8.0.0",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
-    "toad-scheduler": "^1.0.0"
+    "js-yaml": "^4.1.0",
+    "luxon": "^2.4.0",
+    "mariadb": "^3.0.0",
+    "normalize.css": "^8.0.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "toad-scheduler": "^1.6.0"
   },
   "devDependencies": {
-    "@rollup/plugin-yaml": "^3.0.0",
-    "@types/js-yaml": "^4.0.0",
-    "@types/luxon": "^1.0.0",
-    "@types/node": "^16.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@typescript-eslint/eslint-plugin": "^4.0.0",
-    "@typescript-eslint/parser": "^4.0.0",
+    "@rollup/plugin-yaml": "^3.1.0",
+    "@types/js-yaml": "^4.0.5",
+    "@types/luxon": "^2.3.2",
+    "@types/node": "^17.0.35",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.4",
+    "@typescript-eslint/eslint-plugin": "^5.25.0",
+    "@typescript-eslint/parser": "^5.25.0",
     "@vitejs/plugin-react-refresh": "^1.0.0",
-    "chalk": "^2.0.0",
-    "esbuild": "^0.12.0",
-    "esbuild-node-externals": "^1.3.0",
-    "eslint": "^7.0.0",
-    "eslint-plugin-react": "^7.0.0",
-    "glob": "^7.0.0",
-    "sass": "^1.0.0",
-    "vite": "^2.0.0"
+    "chalk": "^5.0.1",
+    "esbuild": "^0.14.39",
+    "esbuild-node-externals": "^1.4.1",
+    "eslint": "^8.16.0",
+    "eslint-plugin-react": "^7.30.0",
+    "glob": "^8.0.3",
+    "sass": "^1.52.1",
+    "vite": "^2.9.9"
   }
 }
