@@ -1,49 +1,34 @@
-import { readFileSync } from 'fs'
-import { Server } from 'https'
-
-import fastify, { FastifyHttpsOptions } from 'fastify'
-import fastifyCors from 'fastify-cors'
-import fastifyCookie from 'fastify-cookie'
-import fastifyRateLimit from 'fastify-rate-limit'
+import fastify from 'fastify'
+import fastifyCors from '@fastify/cors'
+import fastifyCookie from '@fastify/cookie'
+import fastifyRateLimit from '@fastify/rate-limit'
 import fastifySchedule from 'fastify-schedule'
 import c from 'chalk'
 
-import config from './config'
-import State from './utils/state-codes'
-import './database'
+import config from './config.js'
+import State from './utils/state-codes.js'
+import './database.js'
 
-import mainPage from './plugins/main-page'
-import auth from './plugins/auth'
-import shorten from './plugins/shorten'
-import jump from './plugins/jump'
-import manage from './plugins/manage'
-import tasks from './plugins/schedule'
-
-let serverOption: Partial<FastifyHttpsOptions<Server>> = {}
-
-if (config.dev) {
-  serverOption = {
-    https: {
-      key: readFileSync('./ssl/localhost.key'),
-      cert: readFileSync('./ssl/localhost.crt')
-    }
-  }
-}
+import mainPage from './plugins/main-page.js'
+import auth from './plugins/auth.js'
+import shorten from './plugins/shorten.js'
+import jump from './plugins/jump.js'
+import manage from './plugins/manage.js'
+import tasks from './plugins/schedule.js'
 
 const server = fastify({
   logger: {
-    level: config.logger
+    level: config.logger,
   },
-  ...serverOption
 })
 
 if (config.dev) {
   server.register(fastifyCors, {
     origin: [
-      `https://localhost:${config.devport}`
+      `http://localhost:${config.devport}`,
     ],
     methods: 'GET',
-    credentials: true
+    credentials: true,
   })
 }
 
@@ -51,7 +36,7 @@ server.register(fastifyCookie, { secret: config.secret })
 server.register(fastifyRateLimit, {
   max: 100,
   timeWindow: 1000,
-  errorResponseBuilder: () => State.error(429)
+  errorResponseBuilder: () => State.error(429),
 })
 server.register(fastifySchedule)
 
@@ -65,7 +50,7 @@ server.register(tasks)
 server.setNotFoundHandler({
   preHandler: () => {
     server.rateLimit({ max: 5 })
-  }
+  },
 }, () => {
   return State.error(404)
 })
@@ -73,7 +58,7 @@ server.setNotFoundHandler({
 const main = async () => {
   try {
     server.listen(config.port)
-    console.log(`Server is listening at ${c.cyan(`https://localhost:${config.port}`)} ...`)
+    console.log(`Server is listening at ${c.cyan(`http://localhost:${config.port}`)} ...`)
   } catch (err) {
     server.log.error(err)
     process.exit(1)
