@@ -7,11 +7,11 @@ import { resolve } from 'path'
 
 import { FastifyPluginAsync } from 'fastify'
 import fp from 'fastify-plugin'
-import fastifyStatic from 'fastify-static'
+import fastifyStatic from '@fastify/static'
 
 const mainPage: FastifyPluginAsync = async (server) => {
   server.register(fastifyStatic, {
-    root: resolve('./frontend/dist/')
+    root: resolve('./frontend/dist/'),
   })
 
   server.get('/', async (_, reply) => {
