@@ -9,7 +9,7 @@ ReactDOM.render(
   <React.StrictMode>
     <AppContainer />
   </React.StrictMode>,
-  document.getElementById('root')
+  document.getElementById('root'),
 )
 
 document.title = config.title
